+++ common/src/main/java/org/apache/rocketmq/common/message/MessageClientIDSetter.java
@@ -40,7 +40,7 @@ public class MessageClientIDSetter {
         LEN = ip.length + 2 + 4 + 4 + 2;
         ByteBuffer tempBuffer = ByteBuffer.allocate(ip.length + 2 + 4);
         tempBuffer.put(ip);
-        tempBuffer.putShort(UtilAll.getPid());
+        tempBuffer.putShort((short) UtilAll.getPid());
         tempBuffer.putInt(MessageClientIDSetter.class.getClassLoader().hashCode());
         FIX_STRING = UtilAll.bytes2string(tempBuffer.array());
         setStartTime(System.currentTimeMillis());
